% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_proj_data.R
\name{auto_plot}
\alias{auto_plot}
\title{automatic plots for the projdata}
\usage{
auto_plot(data_sf, out_path, out_name)
}
\arguments{
\item{data_sf}{the data that has to be plotted}

\item{out_path}{path where the plot has to be stored}

\item{out_name}{name of the plot}
}
\value{
save the data to the original location
}
\description{
automatic plots for the projdata
}
\examples{
\dontrun{

}
}
