% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_data_ras.R
\name{build_data_ras}
\alias{build_data_ras}
\title{create xlsx for the project data.}
\usage{
build_data_ras(data_name, data_ras)
}
\arguments{
\item{data_name}{the name of the project}

\item{data_ras}{spatRaster object}
}
\value{
a tibble where you have to set the parameters by hand.
}
\description{
create xlsx for the project data.
}
\examples{
\dontrun{
build_data_ras()
}
}
