% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mcp.R
\name{calc_mcp}
\alias{calc_mcp}
\title{function for creating the mcp of telemetry data}
\usage{
calc_mcp(xy, percent = 100)
}
\arguments{
\item{xy}{sf data}

\item{percent}{percentage of mcp. default 100\%}
}
\value{
mcps for all individuals
}
\description{
function for creating the mcp of telemetry data
}
\examples{
\dontrun{
calc_mcp()
}
}
