% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_cols.R
\name{change_cols}
\alias{change_cols}
\title{change columns of a dataframe.}
\usage{
change_cols(path)
}
\arguments{
\item{path}{The path where the data lays}
}
\value{
save the data to the original location
}
\description{
change columns of a dataframe.
}
\examples{
\dontrun{
change_cols(path = ".")
}
}
