% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_proj_data.R
\name{fun_dl_ger_bor}
\alias{fun_dl_ger_bor}
\title{download data from the bkg database}
\usage{
fun_dl_ger_bor()
}
\value{
a shapefile of germany
}
\description{
download data from the bkg database
}
\examples{
\dontrun{
fun_dl_ger_bor()
}
}
