% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_data_files.R, R/build_data_ras.R
\name{fun_epsg}
\alias{fun_epsg}
\title{function for epsg code}
\usage{
fun_epsg()

fun_epsg()
}
\value{
choose an epsg code or add a costum

choose an epsg code or add a costum
}
\description{
function for epsg code

function for epsg code
}
\examples{
\dontrun{
fun_epsg()
}
\dontrun{
fun_epsg()
}
}
