% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_proj_data.R
\name{fun_inter_data_ger}
\alias{fun_inter_data_ger}
\title{dintersect point data with one of the germany shapefiles}
\usage{
fun_inter_data_ger(pts_data)
}
\arguments{
\item{pts_data}{the point data}
}
\value{
a shapefile with part of germany
}
\description{
dintersect point data with one of the germany shapefiles
}
\examples{
\dontrun{
fun_inter_data_ger()
}
}
