% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_raw_data.R
\name{fun_move_raw}
\alias{fun_move_raw}
\title{move raw data from any directory to the _archive folder.}
\usage{
fun_move_raw(data_name, data_path)
}
\arguments{
\item{data_name}{column of the taxon name}

\item{data_path}{complete path to the raw data with the file itself}
}
\description{
move raw data from any directory to the _archive folder.
}
\examples{
\dontrun{
fun_move_raw(taxon_col = "taxon_col", data_path = "path")
}
}
