% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_data_files.R, R/build_data_ras.R
\name{fun_num}
\alias{fun_num}
\title{function for numeric}
\usage{
fun_num(x)

fun_num(x)
}
\arguments{
\item{x}{input}
}
\value{
returns a number

returns a number
}
\description{
function for numeric

function for numeric
}
\examples{
\dontrun{
fun_date("1")
}
\dontrun{
fun_date("1")
}
}
