% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_info.R
\name{get_meta_info}
\alias{get_meta_info}
\title{get meta info for overview plot}
\usage{
get_meta_info(path)
}
\arguments{
\item{path}{where to find the xlsx meta file}
}
\value{
a text file for ggplot
}
\description{
get meta info for overview plot
}
\examples{
\dontrun{

}
}
