% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layers.R
\name{map_layers}
\alias{map_layers}
\title{function for creating an interactive leaflet plot. works for vector data only}
\usage{
map_layers(x, id_col)
}
\arguments{
\item{x}{data}

\item{id_col}{the name of the id column}
}
\value{
an interactive leaflet plot where you can click on the different ids to show
}
\description{
function for creating an interactive leaflet plot. works for vector data only
}
\examples{
\dontrun{
map_layers()
}
}
