% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_data_name.R
\name{proj_data_name}
\alias{proj_data_name}
\title{create name of proj data files}
\usage{
proj_data_name()
}
\value{
a data name in a given format
}
\description{
create name of proj data files
}
\examples{
\dontrun{
proj_data_name()
}
}
