% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_projdata_ras.R
\name{render_projdata_ras}
\alias{render_projdata_ras}
\title{function for rendering an markdown to get an html of the project data}
\usage{
render_projdata_ras(path, data_name, data_ras, proj_data)
}
\arguments{
\item{path}{The path where the data has to be stored}

\item{data_name}{The name of the data}

\item{data_ras}{spatRaster object}

\item{proj_data}{meta data file}
}
\value{
a tibble where you have to set the parameters by hand.
}
\description{
function for rendering an markdown to get an html of the project data
}
\examples{
\dontrun{
render_projdata_ras()
}
}
