% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_raw_data.R
\name{std_raw_data}
\alias{std_raw_data}
\title{standardise the raw data.}
\usage{
std_raw_data(taxon_col, data_path)
}
\arguments{
\item{taxon_col}{name of the taxon}

\item{data_path}{The path where the data has to be stored}
}
\value{
a tibble where you have to set the parameters by hand.
}
\description{
standardise the raw data.
}
\examples{
\dontrun{
build_data_files()
}
}
