% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_data_files.R
\name{build_data_files}
\alias{build_data_files}
\title{create xlsx for the project data.}
\usage{
build_data_files(path, data_name)
}
\arguments{
\item{path}{The path where the data has to be stored}

\item{data_name}{the name of the project}
}
\value{
a tibble where you have to set the parameters by hand.
}
\description{
create xlsx for the project data.
}
\examples{
\dontrun{
build_data_files()
}
}
