% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_projdata.R
\name{render_projdata}
\alias{render_projdata}
\title{function for rendering an markdown to get an html of the project data}
\usage{
render_projdata(path, data_name, data_sf, data_plot, proj_data, id_col, data)
}
\arguments{
\item{path}{The path where the data has to be stored}

\item{data_name}{The name of the data}

\item{data_sf}{sf object}

\item{data_plot}{data for plotting}

\item{proj_data}{meta data file}

\item{id_col}{id column for plotting}

\item{data}{data for table}
}
\value{
a tibble where you have to set the parameters by hand.
}
\description{
function for rendering an markdown to get an html of the project data
}
\examples{
\dontrun{
render_projdata()
}
}
